change floats to doubles in unimodal ternary search

+++ src/main/kotlin/com/falcinspire/hacklib/search/Precision.kt
@@ -1,3 +1,3 @@
 package com.falcinspire.hacklib.search
 
-class Precision(val value: Float)
+class Precision(val value: Double)
